new structure for statistics; in the statistic folder we can group statistic methods as specific services, e.g. developmentFocus
